Revert "feat: TimetableSelection Activity ErrorBoundary 및 Suspense 처리"
This reverts commit 274b1954a3d02c353a26bb567f26774d78bf0683.
