+++ package.json
@@ -31,8 +31,8 @@
     "lucide-react": "^0.473.0",
     "mixpanel-browser": "^2.60.0",
     "motion": "^12.0.6",
-    "react": "^18.3.0",
-    "react-dom": "^18.3.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-use-measure": "^2.1.7",
     "tailwind-merge": "^3.0.1",
     "tailwindcss": "^4.0.0",
@@ -43,8 +43,8 @@
     "@eslint/js": "^9.17.0",
     "@types/lodash": "^4.17.15",
     "@types/mixpanel-browser": "^2.51.0",
-    "@types/react": "^18.3.0",
-    "@types/react-dom": "^18.3.0",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "^18.3.5",
     "@vitejs/plugin-react": "^4.3.4",
     "eslint": "^9.17.0",
     "eslint-config-prettier": "^10.0.1",
