+++ src/components/SoongptErrorBoundary.tsx
@@ -7,10 +7,10 @@ import Wrench from '../assets/wrench.svg';
 
 interface SoongptErrorBoundaryProps {
   children: ReactNode;
-  clientErrorComponent?: ReactElement;
+  FallbackComponent: ReactElement;
 }
 
-const SoongptErrorBoundary = ({ clientErrorComponent, children }: SoongptErrorBoundaryProps) => {
+const SoongptErrorBoundary = ({ FallbackComponent, children }: SoongptErrorBoundaryProps) => {
   return (
     <QueryErrorResetBoundary>
       {({ reset }) => (
@@ -18,43 +18,40 @@ const SoongptErrorBoundary = ({ clientErrorComponent, children }: SoongptErrorBo
           onError={reset}
           fallback={({ error, resetError }) => {
             const errorRange = Math.floor(((error as SoongptError).status ?? 500) / 100);
-            const serverErrorComponent = (
-              <motion.div
-                className="mt-6 flex flex-1 flex-col items-center gap-6 overflow-auto"
-                initial={{ y: 20, opacity: 0 }}
-                animate={{ y: 0, opacity: 1 }}
-                transition={{ duration: 0.3, ease: 'easeOut' }}
-              >
-                <div className="flex w-full flex-col items-center overflow-auto">
-                  <h2 className="text-center text-[28px]/[normal] font-semibold whitespace-pre-wrap">
-                    {'서버와의 연결이\n원활하지 않아요.'}
-                  </h2>
-                  <span className="items mt-1 font-light">
-                    재요청 하더라도 사용자님의 정보는 기억할게요!
-                  </span>
-                </div>
-                <div className="flex flex-1 items-center justify-center">
-                  <img className="my-auto" src={Wrench} alt={'wrench'} />
-                </div>
-                <div className="flex w-full items-center justify-center gap-3">
-                  <button
-                    onClick={resetError}
-                    type="button"
-                    className="bg-primary max-w-52 flex-1 rounded-2xl py-3.5 font-semibold text-white"
-                  >
-                    재요청 할래요
-                  </button>
-                </div>
-              </motion.div>
-            );
-
-            if (clientErrorComponent === undefined) return serverErrorComponent;
 
             switch (errorRange) {
               case 4:
-                return clientErrorComponent;
+                return FallbackComponent;
               default:
-                return serverErrorComponent;
+                return (
+                  <motion.div
+                    className="flex flex-1 flex-col items-center gap-6 overflow-auto"
+                    initial={{ y: 20, opacity: 0 }}
+                    animate={{ y: 0, opacity: 1 }}
+                    transition={{ duration: 0.3, ease: 'easeOut' }}
+                  >
+                    <div className="flex w-full flex-col items-center overflow-auto">
+                      <h2 className="text-center text-[28px]/[normal] font-semibold whitespace-pre-wrap">
+                        {'서버와의 연결이\n원활하지 않아요.'}
+                      </h2>
+                      <span className="items mt-1 font-light">
+                        재요청 하더라도 사용자님의 정보는 기억할게요!
+                      </span>
+                    </div>
+                    <div className="flex flex-1 items-center justify-center">
+                      <img className="my-auto" src={Wrench} alt={'wrench'} />
+                    </div>
+                    <div className="flex w-full items-center justify-center gap-3">
+                      <button
+                        onClick={resetError}
+                        type="button"
+                        className="bg-primary max-w-52 flex-1 rounded-2xl py-3.5 font-semibold text-white"
+                      >
+                        재요청 할래요
+                      </button>
+                    </div>
+                  </motion.div>
+                );
             }
           }}
         >
