+++ src/hooks/usePostTimetable.ts
@@ -4,8 +4,6 @@ import api from '../api/client';
 import { StudentTimetable } from '../schemas/studentSchema';
 import { timetableArrayResponseSchema } from '../schemas/timetableSchema';
 import { transformError } from '../utils/error.ts';
-import { Mixpanel } from '../utils/mixpanel.ts';
-import { SoongptError } from '../schemas/errorSchema.ts';
 
 export const usePostTimetable = () => {
   return useMutation({
@@ -22,9 +20,6 @@ export const usePostTimetable = () => {
         });
 
       return timetableArrayResponseSchema.parse(response);
-    },
-    onError: (error: SoongptError) => {
-      Mixpanel.trackTimetableSelectionError(error);
     },
   });
 };
